+++ fastai/core.py
@@ -283,3 +283,11 @@ def parallel(func, arr:Collection, max_workers:int=None):
             futures = [ex.submit(func,o,i) for i,o in enumerate(arr)]
             for f in progress_bar(concurrent.futures.as_completed(futures), total=len(arr)): pass
 
+def subplots(rows:int, cols:int, imgsize:int=4, figsize:Optional[Tuple[int,int]]=None, title=None, **kwargs):
+    "Like `plt.subplots` but with consistent axs shape, `kwargs` passed to `fig.suptitle` with `title`"
+    figsize = ifnone(figsize, (imgsize*cols, imgsize*rows))
+    fig, axs = plt.subplots(rows,cols,figsize=figsize)
+    if (rows==1 and cols!=1) or (cols==1 and rows!=1): axs = [axs]
+    fig.suptitle(title, **kwargs)
+    return array(axs)
+