+++ CHANGES.md
@@ -20,6 +20,8 @@ of that change.
 
 - `ImageItemList` now takes optional `convert_mode`
 - `Image.show` now uses `defaults.cmap` if no `cmap` passed
+- `bn` param in `conv_layer` replaced by `norm_type` which takes `NormType` enum
+- unet kwargs are passed down to `conv_layer`
 
 ### Fixed: